+++ src/component/App/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 // import logo from './logo.svg';
 // import './App.css';
-import Header from './component/Header';
-import Footer from './component/Footer';
+import Header from '../Header';
+import Footer from '../Footer';
 
 
 function App() {
