+++ src/component/Footer/Copyright.jsx
@@ -4,8 +4,8 @@ const Copyright = () => (
     <div>
         <ul>
             <li>Copyright 2020</li>
-            <li><a>Terms</a></li>
-            <li><a>Privacy</a></li>
+            <li><a href="www.google.com">Terms</a></li>
+            <li><a href="www.google.com">Privacy</a></li>
         </ul>
 
     </div>
