--- src/component/Header/Navbar.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const Navbar = () => (
-    <nav className="navbar">
-        <ul className="navbar__list">
-            <li><a href="#" className="navbar__link">Home</a></li>
-            <li><a href="#" className="navbar__link">Service</a></li>
-            <li><a href="#" className="navbar__link">Log In</a></li>
-        </ul>
-    </nav>
-);
-
-export default Navbar;
