+++ src/component/Header/component/Navbar.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Navbar = () => (
+    <nav className="navbar">
+        <ul className="navbar__list">
+            <li><a href="www.google.com" className="navbar__link">Home</a></li>
+            <li><a href="www.google.com" className="navbar__link">Service</a></li>
+            <li><a href="www.google.com" className="navbar__link">Log In</a></li>
+        </ul>
+    </nav>
+);
+
+export default Navbar;
